+++ src/app/zisseki-demo/[year]/[week]/memo/architecture-diagram.md
